services/backend/requirements.txt
ollama
sentence-transformers
faiss-cpu
flask
numpy